iss643 feature_flag: Fehlende Prüfung vor foreach (PHP-Warning)
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/feature_flag/trunk@83807 735d13b6-9817-0410-8766-e36946ffe9aa

+++ Classes/System/Typo3/TCA.php
@@ -158,6 +158,9 @@ class Tx_FeatureFlag_System_Typo3_TCA
             return;
         }
         $mappings = $this->getMappingRepository()->findAllByForeignTableNameAndUid($id, $table);
+        if (FALSE === is_array($mappings) && FALSE === ($mappings instanceof Tx_Extbase_Persistence_QueryResultInterface)) {
+            return;
+        }
         foreach ($mappings as $mapping) {
             if ($mapping instanceof Tx_FeatureFlag_Domain_Model_Mapping) {
                 $this->getMappingRepository()->remove($mapping);
